refactor tracing into directory (non-breaking) and add new requests wrapper for unsupported services; todo: add to docs
