+++ setup.py
@@ -31,8 +31,10 @@ if __name__ == "__main__":
             'wheel'
         ],
         tests_require=[
-            'pytest',
-            'pytest-runner',
-            'flake8',
-            'pytest-cov', ],
+            'pytest==3.6.2',
+            'pytest-runner==4.2',
+            'flake8==3.5.0',
+            'pycodestyle<2.4.0',
+            'pytest-cov==2.5.1',
+        ],
     )
